\name{moultonTools-package}
\Rdversion{1.1}
\alias{moultonTools-package}
\alias{moultonTools}
\docType{package}
\title{
moultonTools
}
\description{
Tools for applying the Moulton correction for clustering. Primarily based on the book "Mostly harmless econometrics: An empiricists companion", Angrist J. and Pishke J.S., Princeton, Oxford, 2009.
}
\details{
\tabular{ll}{
Package: \tab moultonTools\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2017-03-17\cr
LazyLoad: \tab yes\cr
}

}
\author{
Marina Braschler <marina.braschler@gmail.com>\cr
Thomas Braschler <thomas.braschler@gmail.com>\cr
}

\keyword{misc}

