\name{moultonTools-package}
\Rdversion{1.1}
\alias{moultonTools-package}
\alias{moultonTools}
\docType{package}
\title{
moultonTools
}
\description{
Tools for applying the Moulton correction for clustering. The logics for Moulton factor and intraclass correlation coefficient calculations are primarily based on the book "Mostly harmless econometrics: An empiricists companion", Angrist J. and Pishke J.S., Princeton, Oxford, 2008. The package contains additional convenience functions to perform Moulton-corrected tests in common settings: The functions \link{anova_moulton}, \link{moulton.t.test}, \link{prop.trend.test.moulton} are analogs of the corresponding native R functions with built-in Moulton correction.  
}
\details{
\tabular{ll}{
Package: \tab moultonTools\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2017-03-17\cr
LazyLoad: \tab yes\cr
}

}
\author{
Marina Braschler <marina.braschler@gmail.com>\cr
Thomas Braschler <thomas.braschler@gmail.com>\cr
}

\keyword{misc}

\references{

"Mostly harmless econometrics: An Empiricist's Companion", Angrist J.D., Pischke, J.S., 2008

}



